requirements.txt
requests
beautifulsoup4
pandas
yfinance
quandl
python-dotenv